Fix test_change_hash_type to be more thorough
When changing the hash type, a login should be attempted twice. On the
first attempt, Flask-Security should recognize that the hash algorithm has
changed and update the datastore. On the second attempt, Flask-Security
should properly validate against the updated datastore.

This commit causes the test to correctly fail. The underlying behavior
will be fixed in a subsequent commit.