Add a new test for changing a hash to plaintext
In addition to the existing test case which tested the change from
plaintext to hashed passwords, this adds an addition test for the change
from hashed to plaintext passwords.

Either direction of change should work, provided that the salt is known
at all times, or at least known when first changing to hashed passwords.

This new test correctly fails. The underlying behavior will be fixed in a
subsequent commit.