Fix verify_and_update_password to work when changing to or from plaintext
If password is being updated to change from plaintext to encrypted, it
must be verified while non-salted, then encrypted while salted. As the
salting is done outside of Passlib, this case must be explicitly handled.

The same logic applies when changing from encrypted to plaintext passwords.
As the password is being provided to verify_and_update_password in
plaintext, it is possible, though not encouraged, to make the transition
in this direction too.