Encrypt raw password, not HMAC of password.
This passes passwords directly to passlib's CryptContext
methods and relies on it to do the hard work.

For backwards compatibility we fall back to trying the old HMAC approach
when verifying passwords, but only if a salt has been set.

Previously, we passed the HMAC'd password to passlib, which meant that
we weren't compatible with other users of passlib (such as Django) or
the Modular Crypt Format at
https://pythonhosted.org/passlib/modular_crypt_format.html

Fixes #320