datastore: notify ORM when modifying User.active
Without this, the following:

    user = datastore.get_user(...)
    datastore.toggle_active(user)
    datastore.commit()

did not commit the change to user.active with the Mongo or Peewee
backends.

I think it worked with the SQLAlchemy backend because objects returned
from a query are already part of the session, and
SQLAlchemyDatastore.commit() actually does something; whereas the Mongo
and Peewee stores do nothing in .commit() and everything in .put().