Upgrade passlib, remove bcrypt restriction
Due to a bug in passlib [0], use of bcrypt>=0.2.0 was not working
properly. As of passlib==1.6.4, that bug is fixed [1]. The minimum
required version of passlib is upgraded to 1.6.4 to address this issue
and properly support current versions of bcrypt. Because pypy is
outdated on travis [2,3], a fix [4] to bring it up to date is needed
(locally, tests pass with pypy >= 2.6 but not 2.5, the current version
used on travis).

[0] https://bitbucket.org/ecollins/passlib/issues/56
[1] https://bitbucket.org/ecollins/passlib/src/fa26802fc88fbc6b2d4b9676970bed7e6d53fc9d/CHANGES?at=default&fileviewer=file-view-default#CHANGES-130,131
[2] https://github.com/travis-ci/travis-ci/issues/4756
[3] https://github.com/travis-ci/travis-ci/issues/5027
[4] https://github.com/travis-ci/travis-ci/issues/5027#issuecomment-170939193