views: allow adding a csrf token to json responses
flask-wtf's CsrfProtect makes it hard to use flask-security with json
requests because a csrf token is needed when posting json data.

This difficulty can be alleviated by teaching flask-security
provide a csrf token in its responses.

Teach _render_json() to supply a csrf_token value in the response when
the SECURITY_JSON_INCLUDE_CSRF configuration is True.  The presence
of csrf_token allows clients to respond back with the X-CSRFToken
header set to its value.  This allows flask-security json
requests to work without needing to disable csrf.

Related-to: #421
Signed-off-by: David Aguilar <davvid@gmail.com>