Support passwordless JSON logins
Only the first phase of the passwordless login supported JSON based requests.
The login itself, usually done by clicking on the link in the login email,
supported only html requests and always responded with a `Set-Cookie` header.
Now the login itself can be sent with `application/json` in the `Accept` header
resulting in a JSON response that contains `authorization_token` in the
`user` attribute contained in the `response` attribute.